refactor(scheduler): share fifo scheduler common logic
